// src/enums/contract-type.enum.ts
export enum enumContractType {
  CLASSIC_MINT = "ClassicMint",
  CLASSIC_MINT_WITH_WL = "WLClassicMint",
  DUTCH_AUCTION = "DutchAuction",
  DUTCH_AUCTION_WITH_WL = "WLDutchAuction",
  FAIR_DUTCH_AUCTION = "FairDutchAuction",
  FAIR_DUTCH_AUCTION_WITH_WL = "WLFairDutchAuction",
  PURE_WHITELIST = "PureWhitelist",
}